balance
removed all rifles & shotguns from start, dropped a soviet pistol as well.
On completion of first NF mission, shotgun & krag will be unlocked.

There is now a desk at the start of the mission which has a few rifles and what not on it to start the boys off. Everyone else has to use a pistol.